+++ module/Formation/src/Formation/Service/Notification/FormationNotificationFactory.php
@@ -2,7 +2,9 @@
 
 namespace Formation\Service\Notification;
 
+use Formation\Entity\Db\Formateur;
 use Formation\Entity\Db\Inscription;
+use Formation\Entity\Db\Session;
 use Formation\Provider\Template\MailTemplates;
 use Notification\Exception\RuntimeException;
 use Notification\Factory\NotificationFactory;
@@ -162,6 +164,35 @@ class FormationNotificationFactory extends NotificationFactory
         $notif = new Notification();
         $notif
             ->setTo($mail)
+            ->setSubject($rendu->getSujet())
+            ->setBody($rendu->getCorps())
+        ;
+
+        return $notif;
+    }
+
+    public function createNotificationSessionImminenteFormateur(Session $session) : Notification
+    {
+        $vars = [
+            'formation' => $session->getFormation(),
+            'session'   => $session,
+        ];
+        $rendu = $this->getRenduService()->generateRenduByTemplateCode(MailTemplates::SESSION_IMMINENTE_FORMATEUR, $vars);
+
+        $mails = [];
+        /** @var Formateur $formateur */
+        foreach ($session->getFormateurs() as $formateur) {
+            $mail = $formateur->getIndividu()->getEmailUtilisateur()??$formateur->getIndividu()->getEmailPro();
+            if ($mail !== null) $mails[] = $mail;
+        }
+
+        if (empty($mails)) {
+            throw new RuntimeException("Aucune adresse mail trouvée pour les formateurs de la session {$session->getCode()}.");
+        }
+
+        $notif = new Notification();
+        $notif
+            ->setTo($mails)
             ->setSubject($rendu->getSujet())
             ->setBody($rendu->getCorps())
         ;